refactor(core): share durable file lease
